리뷰 반영 : 코드의 가독성을 더 높게 변경
